+++ src/test/java/org/jolokia/docker/maven/service/RunServiceTest.java
@@ -2,16 +2,29 @@ package org.jolokia.docker.maven.service;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Properties;
+
+import org.apache.commons.io.IOUtils;
+import org.apache.maven.project.MavenProject;
+import org.jolokia.docker.maven.access.ContainerCreateConfig;
+import org.jolokia.docker.maven.access.ContainerHostConfig;
+import org.jolokia.docker.maven.access.DockerAccess;
+import org.jolokia.docker.maven.access.PortMapping;
+import org.jolokia.docker.maven.config.ImageConfiguration;
+import org.jolokia.docker.maven.config.RestartPolicy;
+import org.jolokia.docker.maven.config.RunImageConfiguration;
+import org.jolokia.docker.maven.config.VolumeConfiguration;
+import org.jolokia.docker.maven.util.Logger;
+import org.junit.Test;
+import org.skyscreamer.jsonassert.JSONAssert;
 
 import mockit.Expectations;
 import mockit.Mocked;
-import org.apache.commons.io.IOUtils;
-import org.apache.maven.project.MavenProject;
-import org.jolokia.docker.maven.access.*;
-import org.jolokia.docker.maven.config.*;
-import org.junit.Test;
-import org.skyscreamer.jsonassert.JSONAssert;
 
 
 /**
@@ -23,6 +36,12 @@ public class RunServiceTest {
     @Mocked
     private DockerAccess docker;
 
+    @Mocked
+    private Logger log;
+    
+    @Mocked
+    private QueryService queryService;
+    
     @Mocked
     private MavenProject mavenProject;
 
@@ -63,46 +82,60 @@ public class RunServiceTest {
                         .restartPolicy(restartPolicy())
                         .build();
 
-        RunService runService = new RunService();
-        PortMapping portMapping = runService.getPortMapping(runConfig, new Properties());
-
         new Expectations() {{
-            docker.getContainerName((String) withNotNull());
-            result = "redis";
-            minTimes = 1;
+                queryService.getContainerName("redisContainer");
+                result = "redis";
 
+                queryService.getContainerName("parentContainer");
+                result = "parentContainer";
+
+                queryService.getContainerName("otherContainer");
+                result = "otherContainer";
+
+                queryService.getContainerName("redisContainer");
+                result = "redis";
+                queryService.getContainerName("parentContainer");
+                result = "parentContainer";
+
+                queryService.getContainerName("otherContainer");
+                result = "otherContainer";
         }};
+        
+        ContainerTracker tracker = new ContainerTracker();
+        RunService runService = new RunService(docker, queryService, tracker, log);
+        
+        PortMapping portMapping = runService.getPortMapping(runConfig, new Properties());
 
         // Better than poking into the private vars would be to use createAndStart() with the mock to build up the map.
         ImageConfiguration imageConfig2 = new ImageConfiguration.Builder().alias("db").name("redis3").build();
-        putToPrivateMap(runService, "containerImageNameMap", imageConfig2.getName(), "redisContainer");
+        putToPrivateMap(tracker, "containerImageNameMap", imageConfig2.getName(), "redisContainer");
         if (imageConfig2.getAlias() != null) {
-            putToPrivateMap(runService,"imageAliasMap",imageConfig2.getAlias(), imageConfig2.getName());
+            putToPrivateMap(tracker,"imageAliasMap",imageConfig2.getAlias(), imageConfig2.getName());
         }
         ImageConfiguration imageConfig1 = new ImageConfiguration.Builder().alias("parent").name("parentName").build();
-        putToPrivateMap(runService,"containerImageNameMap",imageConfig1.getName(), "parentContainer");
+        putToPrivateMap(tracker,"containerImageNameMap",imageConfig1.getName(), "parentContainer");
         if (imageConfig1.getAlias() != null) {
-            putToPrivateMap(runService,"imageAliasMap",imageConfig1.getAlias(), imageConfig1.getName());
+            putToPrivateMap(tracker,"imageAliasMap",imageConfig1.getAlias(), imageConfig1.getName());
         }
         ImageConfiguration imageConfig = new ImageConfiguration.Builder().alias("other:ro").name("otherName").build();
-        putToPrivateMap(runService,"containerImageNameMap",imageConfig.getName(), "otherContainer");
+        putToPrivateMap(tracker,"containerImageNameMap",imageConfig.getName(), "otherContainer");
         if (imageConfig.getAlias() != null) {
-            putToPrivateMap(runService,"imageAliasMap", imageConfig.getAlias(), imageConfig.getName());
+            putToPrivateMap(tracker,"imageAliasMap", imageConfig.getAlias(), imageConfig.getName());
         }
-        ContainerCreateConfig containerConfig = runService.createContainerConfig(docker, "base", runConfig, portMapping, new Properties());
+        ContainerCreateConfig containerConfig = runService.createContainerConfig("base", runConfig, portMapping, new Properties());
 
         String expectedConfig = loadFile("docker/containerCreateConfigAll.json");
         JSONAssert.assertEquals(expectedConfig, containerConfig.toJson(), true);
 
-        ContainerHostConfig startConfig = runService.createContainerHostConfig(docker, runConfig, portMapping);
+        ContainerHostConfig startConfig = runService.createContainerHostConfig(runConfig, portMapping);
         String expectedHostConfig = loadFile("docker/containerHostConfigAll.json");
         JSONAssert.assertEquals(expectedHostConfig, startConfig.toJson(), true);
     }
 
-    private void putToPrivateMap(RunService runService, String varName, String key, String value) throws NoSuchFieldException, IllegalAccessException {
-        Field field = runService.getClass().getDeclaredField(varName);
+    private void putToPrivateMap(ContainerTracker tracker, String varName, String key, String value) throws NoSuchFieldException, IllegalAccessException {
+        Field field = tracker.getClass().getDeclaredField(varName);
         field.setAccessible(true);
-        Map<String,String> map = (Map<String, String>) field.get(runService);
+        Map<String,String> map = (Map<String, String>) field.get(tracker);
         map.put(key,value);
     }